+++ tiledb/cf/netcdf_engine/api.py
@@ -31,7 +31,6 @@ def from_netcdf(
     offsets_filters: Optional[tiledb.FilterList] = None,
     attrs_filters: Optional[tiledb.FilterList] = None,
     copy_metadata: bool = True,
-    use_virtual_groups: bool = False,
 ):
     """Converts a NetCDF input file to nested TileDB CF dataspaces.
 
@@ -71,10 +70,6 @@ def from_netcdf(
         attrs_filters: Default filters for all attributes.
         copy_metadata: If  ``True`` copy NetCDF group and variable attributes to
             TileDB metadata. If ``False`` do not copy metadata.
-        use_virtual_groups: If ``True``, create a virtual group using ``output_uri``
-            as the name for the group metadata array. All other arrays will be named
-            using the convention ``{uri}_{array_name}`` where ``array_name`` is the
-            name of the array.
     """
     from .converter import NetCDF4ConverterEngine, open_netcdf_group
 
@@ -98,28 +93,14 @@ def from_netcdf(
             offsets_filters=offsets_filters,
             attrs_filters=attrs_filters,
         )
-        if use_virtual_groups:
-            group_uri = (
-                output_uri
-                if netcdf_group.path == "/"
-                else output_uri + netcdf_group.path.replace("/", "_")
-            )
-            converter.convert_to_virtual_group(
-                group_uri,
-                output_key,
-                output_ctx,
-                input_netcdf_group=netcdf_group,
-                copy_metadata=copy_metadata,
-            )
-        else:
-            group_uri = output_uri + netcdf_group.path
-            converter.convert_to_group(
-                group_uri,
-                output_key,
-                output_ctx,
-                input_netcdf_group=netcdf_group,
-                copy_metadata=copy_metadata,
-            )
+        group_uri = output_uri + netcdf_group.path
+        converter.convert_to_group(
+            group_uri,
+            output_key,
+            output_ctx,
+            input_netcdf_group=netcdf_group,
+            copy_metadata=copy_metadata,
+        )
         if recursive:
             for subgroup in netcdf_group.groups.values():
                 recursive_convert(subgroup)