+++ HISTORY.md
@@ -8,6 +8,8 @@
 
 ### Deprecation
 
+* Removes virtual groups.
+
 ### Bug Fixes
 
 ### New Features